数字课程售卖系统功能基本完成

+++ core/src/main/java/cn/sinjinsong/eshop/core/properties/OrderProperties.java
@@ -4,5 +4,5 @@ package cn.sinjinsong.eshop.core.properties;
  * Created by SinjinSong on 2017/10/6.
  */
 public class OrderProperties {
-    public static final Integer TIME_OUT_SPAN = 5;//5min
+    public static final long TIME_OUT_SPAN = 5L;//5min
 }
